frontend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.7.2
    # via
    #   channels
    #   django
blinker==1.6.2
    # via flask
boto3==1.26.155
    # via -r requirements.in
botocore==1.29.155
    # via
    #   boto3
    #   s3transfer
cacheout==0.14.1
    # via -r requirements.in
cachetools==5.3.1
    # via google-auth
certifi==2023.5.7
    # via requests
channels==4.0.0
    # via django-plotly-dash
charset-normalizer==3.1.0
    # via requests
click==8.1.3
    # via flask
dash==2.9.3
    # via
    #   -r requirements.in
    #   dash-bootstrap-components
    #   django-plotly-dash
dash-bootstrap-components==1.4.1
    # via
    #   -r requirements.in
    #   django-plotly-dash
dash-breakpoints==0.1.0
    # via -r requirements.in
dash-core-components==2.0.0
    # via dash
dash-html-components==2.0.0
    # via dash
dash-table==5.0.0
    # via dash
django==4.2.2
    # via
    #   -r requirements.in
    #   channels
    #   django-plotly-dash
    #   django-storages
django-environ==0.10.0
    # via -r requirements.in
django-plotly-dash==2.2.1
    # via -r requirements.in
django-storages[google]==1.13.2
    # via -r requirements.in
dpd-components==0.1.0
    # via django-plotly-dash
flask==2.3.2
    # via
    #   dash
    #   django-plotly-dash
google-api-core[grpc]==2.11.1
    # via
    #   google-cloud-appengine-logging
    #   google-cloud-core
    #   google-cloud-logging
    #   google-cloud-secret-manager
    #   google-cloud-storage
google-auth==2.20.0
    # via
    #   -r requirements.in
    #   google-api-core
    #   google-cloud-core
    #   google-cloud-storage
google-cloud-appengine-logging==1.3.0
    # via google-cloud-logging
google-cloud-audit-log==0.2.5
    # via google-cloud-logging
google-cloud-core==2.3.2
    # via
    #   google-cloud-logging
    #   google-cloud-storage
google-cloud-logging==3.5.0
    # via -r requirements.in
google-cloud-secret-manager==2.16.1
    # via -r requirements.in
google-cloud-storage==2.9.0
    # via
    #   -r requirements.in
    #   django-storages
google-crc32c==1.5.0
    # via google-resumable-media
google-resumable-media==2.5.0
    # via google-cloud-storage
googleapis-common-protos[grpc]==1.59.1
    # via
    #   google-api-core
    #   google-cloud-audit-log
    #   grpc-google-iam-v1
    #   grpcio-status
grpc-google-iam-v1==0.12.6
    # via
    #   google-cloud-logging
    #   google-cloud-secret-manager
grpcio==1.54.2
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
grpcio-status==1.54.2
    # via google-api-core
gunicorn==20.1.0
    # via -r requirements.in
idna==3.4
    # via requests
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
markupsafe==2.1.3
    # via
    #   jinja2
    #   werkzeug
numpy==1.25.0
    # via -r requirements.in
packaging==23.1
    # via plotly
plotly==5.15.0
    # via
    #   dash
    #   django-plotly-dash
proto-plus==1.22.2
    # via
    #   google-cloud-appengine-logging
    #   google-cloud-logging
    #   google-cloud-secret-manager
protobuf==4.23.3
    # via
    #   google-api-core
    #   google-cloud-appengine-logging
    #   google-cloud-audit-log
    #   google-cloud-logging
    #   google-cloud-secret-manager
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   proto-plus
psycopg2-binary==2.9.6
    # via -r requirements.in
pyasn1==0.5.0
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.3.0
    # via google-auth
python-dateutil==2.8.2
    # via
    #   -r requirements.in
    #   botocore
pytz==2023.3
    # via -r requirements.in
requests==2.31.0
    # via
    #   -r requirements.in
    #   google-api-core
    #   google-cloud-storage
rsa==4.9
    # via google-auth
s3transfer==0.6.1
    # via boto3
six==1.16.0
    # via
    #   google-auth
    #   python-dateutil
sqlparse==0.4.4
    # via django
stream-zip==0.0.63
    # via -r requirements.in
tenacity==8.2.2
    # via plotly
urllib3==1.26.16
    # via
    #   botocore
    #   google-auth
    #   requests
werkzeug==2.3.6
    # via
    #   django-plotly-dash
    #   flask

# The following packages are considered to be unsafe in a requirements file:
# setuptools
